feat/room-info-stomp -> develop (#19)
* fix: 데이터 타입 wrapper로 변경

* fix: websocket으로 인한 api 수정

* fix: limit에 맞는 데이터 타입으로 교체

* fix: db erd에 맞게 수정

* feat: send prefix 및 method 출력 로그 작성

* feat: 방 정보 전달 코드 작성

* feat: 세션 관리 service 작성

------
Close #14

+++ src/main/java/com/maps/coin/repository/QuestionRepository.java
@@ -8,5 +8,5 @@ import org.springframework.data.repository.query.Param;
 
 public interface QuestionRepository extends JpaRepository<Question, Long> {
     @Query(value = "SELECT * FROM question order by RANDOM() limit :size * :size", nativeQuery = true)
-    List<Question> findRandomQuestion(@Param(value="size") Long size);
+    List<Question> findRandomQuestion(@Param(value="size") Integer size);
 }
